// types.ts
export interface User {
  id: string;
  name: string;
  email: string;
  password?: string;
  role: 'admin' | 'user' | 'teacher_arabic' | 'teacher_philosophy' | 'teacher_social';
  totalEarnings: number;
  avatar?: string;
  volume?: number;
  theme?: 'light' | 'dark';
  lastLogin?: string; // YYYY-MM-DD
  streak?: number; // Consecutive days logged in
}
// ...
export interface ChatMessage {
  id: number;
  user_id: string;
  content: string; // Will store text or "Voice Message" label
  type: 'text' | 'audio' | 'image'; // Added 'image'
  media_url?: string; // New field for audio/image URL
  created_at: string;
  likes: number;
  subject?: string;
  profiles?: {
    name: string;
    avatar: string;
    role: string;
  }
}